Json deserialization complete and ready to use

+++ WeatherApp/WeatherApp/Models/Main.cs
@@ -8,8 +8,8 @@ namespace WeatherApp.Models
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public float pressure { get; set; }
+        public float humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
