+++ WeatherApp/WeatherApp/Models/WeatherApiModel.cs
@@ -11,13 +11,13 @@ namespace WeatherApp.Models
             public List<Weather> weather { get; set; }
             public string @base { get; set; }
             public Main main { get; set; }
-            public int visibility { get; set; }
+            public float visibility { get; set; }
             public Wind wind { get; set; }
             public Clouds clouds { get; set; }
-            public int dt { get; set; }
+            public float dt { get; set; }
             public Sys sys { get; set; }
             public int id { get; set; }
             public string name { get; set; }
-            public int cod { get; set; }       
+            public float cod { get; set; }       
     }
 }
