+++ WeatherApp/WeatherApp/Models/Wind.cs
@@ -8,6 +8,6 @@ namespace WeatherApp.Models
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public float deg { get; set; }
     }
 }
